requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
flask>=3.0.0
akshare
